Further split VSCodeServer into individual files
